Add login and register input validation middleware and refactor controllers
